[860qg79et] clickable textfield
